Day 19 part 2

+++ day19/day19.h
@@ -28,7 +28,7 @@ ParsedInput parse_input(std::ifstream &input_stream);
 
 int part1(const ParsedInput &input);
 
-int part2(const ParsedInput &input);
+long long part2(const ParsedInput &input);
 
 } // namespace day19
